refactor(DXswapRouter): Join and rename the files with the DXswapRouter tests cases into one.
