+++ test/shared/fixtures.ts
@@ -11,8 +11,7 @@ import ERC20 from '../../build/contracts/ERC20.json'
 import WETH9 from '../../build/contracts/WETH9.json'
 import UniswapV1Exchange from '../../build/contracts/UniswapV1Exchange.json'
 import UniswapV1Factory from '../../build/contracts/UniswapV1Factory.json'
-import DXswapRouter01 from '../../build/contracts/DXswapRouter01.json'
-import DXswapRouter02 from '../../build/contracts/DXswapRouter02.json'
+import DXswapRouter from '../../build/contracts/DXswapRouter.json'
 import DXswapMigrator from '../../build/contracts/DXswapMigrator.json'
 import RouterEventEmitter from '../../build/contracts/RouterEventEmitter.json'
 
@@ -27,8 +26,6 @@ interface V2Fixture {
   WETHPartner: Contract
   factoryV1: Contract
   factoryV2: Contract
-  router01: Contract
-  router02: Contract
   routerEventEmitter: Contract
   router: Contract
   migrator: Contract
@@ -51,15 +48,14 @@ export async function v2Fixture(provider: Web3Provider, [wallet]: Wallet[]): Pro
   // deploy V2
   const factoryV2 = await deployContract(wallet, DXswapFactory, [wallet.address])
 
-  // deploy routers
-  const router01 = await deployContract(wallet, DXswapRouter01, [factoryV2.address, WETH.address], overrides)
-  const router02 = await deployContract(wallet, DXswapRouter02, [factoryV2.address, WETH.address], overrides)
+  // deploy router
+  const router = await deployContract(wallet, DXswapRouter, [factoryV2.address, WETH.address], overrides)
 
   // event emitter for testing
   const routerEventEmitter = await deployContract(wallet, RouterEventEmitter, [])
 
   // deploy migrator
-  const migrator = await deployContract(wallet, DXswapMigrator, [factoryV1.address, router01.address], overrides)
+  const migrator = await deployContract(wallet, DXswapMigrator, [factoryV1.address, router.address], overrides)
 
   // initialize V1
   await factoryV1.createExchange(WETHPartner.address, overrides)
@@ -88,9 +84,7 @@ export async function v2Fixture(provider: Web3Provider, [wallet]: Wallet[]): Pro
     WETHPartner,
     factoryV1,
     factoryV2,
-    router01,
-    router02,
-    router: router02, // the default router, 01 had a minor bug
+    router,
     routerEventEmitter,
     migrator,
     WETHExchangeV1,
